Move functions related to count down in a helper file and fix bug with add/remove number of focus sessions
